requirements.txt
numpy==1.13.3
Flask==0.12.2
Pillow==4.3.0
imagehash==3.7
pymongo==3.5.1
scikit_learn==0.19.1
